(green) rearrange the routes, and make test pass
Cover the new routes with tests in spec/routing/forks_routing_spec.rb
and spec/routing/tables_routing_spec.rb

Many baseline tests were broken by these changes (and many of those
broken tests indicated changes that needed to be made) in the controller
and views for forks, which are always located on tables, we start
enforcing that.

Forks must belong to a table within the RESTful resource hierarchy.

(eg. if a fork should be created, you create it at /tables/1/forks with
a POST.)

Rails scaffold does most of this for free, if you just keep the default
tests green!